[US_AZ] Ingest Major/Minor violation flag as incident metadata  (Recidiviz/recidiviz-data#47093)
## Description of the change

This PR adds a description to one field (`CHARGE_ID`) and a couple of
datetime parsers to the raw data config for `AZ_DOC_DSC_COORD_REVIEW`,
so that `CHARGE_ID` will appear in the latest view and the data will
import.

It also pulls the `CHARGE_ID` into the ingest view for
`incarceration_incidents` so that we can ingest it as incident metadata.
This will let us know easily if a given incarceration incident was
considered "major" or "minor" by the state.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#46794

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 872fd78dae79527eb083f8ef875b1fb8dea1ccb3