[US_AZ] Distinction for reasons fields for different assessments (Recidiviz/recidiviz-data#47213)
## Description of the change

Changed the name of the reasons fields for our intake assessment and
ORAS assessment criteria in order for our TES to distinguish between
which of the reasons fields is related to which assessment (previously,
both criteria had a reasons field of 'assessment level' which was not
distinguishing which assessment level to use based on eligibility. This
change will also allow us to make necessary changes on the frontend to
account for the Admin Supervision criteria that states that anyone with
an ineligible risk score can be considered for eligibility after 15
months of supervision without violations.

This change has been uploaded to sandbox under `shalin` and all
validations are passing.
There are no changes to eligibility because this is just a reasons field
change;

However, we are seeing a significant change in the group of people who
were historically eligible or almost eligible and had a misaligned
assessment level. (i.e. they were showing as eligible but their reasons
field showed an ineligible assessment level, which was happening because
of the above nondeterminism. these people were ineligible for one
assessment but eligible via another one). The number of these cases
drops from 3522 to 4. This means that there are currently (and
accurately!) only 4 people historically and/or in the future who will
meet the 15 months on supervision violation free criteria while having
an ineligible initial risk score AND an ineligible ORAS score. Again,
this change is because of arbitrariness on the assessment level passed
onto the frontend and not any actual logic bug.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c33aef78021428ef87f6b92be329d55f6ae4455c