[US_UT][US_AZ][US_ND] Fix everything! (Recidiviz/recidiviz-data#47291)
## Description of the change

Many errors today. This PR makes the folllowing fixes:
- **North Dakota:** Maps a new raw text value for
`StateIncarcerationPeriodReleaseReason` in ND. Successful sandbox
pipeline, in prod where the new data lives, is
[here](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-09-05_06_30_01-8871017775861843094;graphView=0?project=recidiviz-123&inv=1&invt=Ab1nOg&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))).
- **Utah:** Removes a field from `user_typ_cd` raw data table that only
appeared in a single transfer on 9/3. Loaded this file successfully to a
sandbox in prod after making the change.
- **Arizona:** Pulls only active incarceration incident coordinator
review rows into the `incarceration_incidents` ingest view in AZ so that
there will not be conflicting information in the resulting objects.
Successful sandbox pipeline
[here](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-09-05_06_51_43-16225589718940715432;mainTab=JOB_GRAPH;graphView=0?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22))).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d07e63758193455dda89f18d5810d997b8a6b92e