[US_TN] Correctly close out safekeeping periods even when period marking edge gets dropped (Recidiviz/recidiviz-data#47656)
## Description of the change

In TN, we normalize PFI to SAFEKEEPING if an IP comes after a period
with "SAREC" in the admission reason raw text, up until a period has an
admission reason of NEW_ADMISSION or REVOCATION, in which case we "end
the safekeeping period" and stop overriding the PFI. This usually works
fine, but we recently found a case where someone's safekeeping period
_should_ have ended due to a NEW_ADMISSION admission reason, but the
NEW_ADMISSION period is a zero-day period that gets dropped, so we never
recognized that the person's safekeeping ended and continued overriding
their PFI with SAFEKEEPING.

We haven't been able to find a good way to resolve this zero-day period
problem, so we needed to take a different approach. This change updates
the `infer_additional_periods` step of IP normalization to append a
"-SKEND" flag to the admission_reason_raw_text of any IP that starts on
the same day as an IP with admission reason NEW_ADMISSION or REVOCATION.
This means that even if the period that actually has one of these
admission reasons gets dropped, we can still recognize that the new
admission / revocation occurred based on the admission reason raw text
of the other period(s) starting on the same date. Accordingly, the
`standardize_purpose_for_incarceration_values` normalization step now
checks for this flag in the admission reason raw text, rather than
checking if a period starts on the same date as a period with admission
reason NEW_ADMISSION / REVOCATION.

This change takes advantage of the fact that IP normalization does the
following steps in this order:
1. `infer_additional_periods`
2. `_drop_periods_from_calculations`
3. `standardize_purpose_for_incarceration_values`

The SKEND flags get added in `infer_additional_periods` before zero-day
periods get dropped, and then once the periods do get dropped in
`_drop_periods_from_calculations`, the SKEND flags will still be
available to `standardize_purpose_for_incarceration_values`.
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#9562

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c5fb99077ea2510739a8c913defeec6c60b70fc7