[US_AZ] Split Violations & Legal Status criteria up (Recidiviz/recidiviz-data#47662)
## Description of the change

We had previously combined some criteria because of complexity issues in
running the TES files; however, now with expanded complexity allowance,
we can go back and revert some of the changes we made. This change
splits up our violations and legal status criteria out of the
CriteriaGroup they were previously in together.

This has been uploaded to sandbox, under `shalin` and all validations
are passing.

Something to note: Once this change is deployed, we will need to go into
Admin Panel and make similar naming changes to ensure the side panel
shows the criteria as expected.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#47661

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8ecdaf9e8679fc65bcf401d22eadfc09ae39ad7a