[lineage] add automcomplete search (Recidiviz/recidiviz-data#47608)
## Description of the change

adds autocomplete search in the top left of the screen. and's
autocomplete is annoying limited -- afaik, if you implement your own
`filterOption`, you can't filter options within groups, but you can if
you use their default?? so i am just using their default as i felt like
grouping options by dataset made things much more readable.

when an option is selected/searched and it is a valid node, the url is
updated to provide a permalink and the graph is cleared to just reflect
this node (the view is reset to a sensible starting zoom)

if the option that is searched doesn't exist, we display an error
message and reset the text (as well as doing the same for bad urls)

e.g. ---

https://github.com/user-attachments/assets/02c7ed23-ed29-4b91-bb6b-ac32d1b3bd5e

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of https://github.com/Recidiviz/recidiviz-data/issues/46345

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 051d4d42d9951f4de2221858ec7af880c5cf474a