[US_AZ] Update logic for hydrating is_violent (Recidiviz/recidiviz-data#47725)
## Description of the change

The AZ data team advised us on how they identify when a charge is
violent. This changes the ingest view to use their logic instead of the
`VIOLENCE_FLAG` in the `AZ_DOC_SC_ARS_CODE`, which has proven
unreliable.

These mappings are a bit different, and not necessarily categorical -
they are not based on the charge description, but on whether or not
there was a weapon involved and whether or not someone was injured. That
means that after this change, there will be multiple charges with the
same description but different `is_violent` values. [This
sheet](https://docs.google.com/spreadsheets/d/1YdqYuFNqdFiU8-ht3uENoG7zGTg7OOKI1-tbkv59ZTE/edit?gid=1644126337#gid=1644126337)
shows for each charge description, the number of records where that
charge is violent and not violent according to this logic.

There are some weird entries there. For instance, 1st degree murder does
not always meet the mark to be considered violent by this logic. I wrote
the research team to ask about this.

Overall, 104,619 historical charges went from not violent to violent
after this change.

**Eligibility changes for TPR:** 11 people will become ineligible for
TPR after this change, since one of the criteria requires that you have
never been convicted of a violent offense and this dramatically
increases the number of offenses we flag as violent.
- One of those people has only a drug offense that is being marked as
violent by this logic - I made a ticket to ask the research team about
that [here](https://github.com/Recidiviz/recidiviz-data/issues/48490).

[sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-central1/2025-09-17_13_58_31-8126552430108698747;graphView=1?project=recidiviz-staging&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

[extended validation
notes](https://www.notion.so/recidiviz/Change-logic-for-is_violent-2707889f4d19804b9e81fa52a4c9395f)
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8e3fb291213bc7897abd79c72048e88bbd347b10