[lineage] add stateful filtering (Recidiviz/recidiviz-data#48518)
## Description of the change

this PR adds stateful filtering to the lineage graph -- apologies for
the size of this PR

there ar four parts to this PR --
1. i moved the computeds from autocomplete to a new store called
`UiStore` which i use in future PRs to manage info about things layered
on top of the graph (like filters, autocomplete, detail page, etc). this
could definitely just live on `GraphStore` but it felt nice to keep that
class somewhat small since it's a little complex already
2. i added two new components -- `NodeFilterModal` which is the popup
when you click the filter button next to the autocomplete (i opted for
this as opposed the existing toolbar to make it easier to find) and
`NodeFilterSelect` which is a small wrapper around antd's Select
3. i added a `NodeFilter` class that's implemented in a sort of
roundabout way -- i've never really built a flexible, composite
filtering system before so am very very very open to suggestions if this
feels wrong. as of right now, there are only 2 filter types (include a
state code and exclude a dataset) but i assume we'll add more in the
future based on feedback, so i wanted to keep it slightly flexible to
that end
4. filter application logic -- in the case where the graph is expanded
and we are applying filters, we might end up making certain nodes
"unreachable" or disconnected from the main body of the graph. the
simplest way i thought to solve this was to use the "selected node" (the
one in the url) as our starting node and bfs our way to find all
reachable nodes, marking all nodes as `hidden` that are not reachable

here's a quick walkthrough video

https://github.com/user-attachments/assets/eca08b36-0b59-4cef-9537-901cfc35e6b9

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

n/a

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1cd63b418d10619a1b69f0adfcf22510dcd568fa