[lineage] add downstream search (Recidiviz/recidiviz-data#48537)
## Description of the change

this PR adds downstream search; it works in 3 parts

1. admin panel api -- instead of computing it on the fly, i chose to
cache all of the downstream/upstream dependencies in redis so we just
need to do a simple key lookup. i had do to some additional wrangley
things to get source table behavior to match up to what we want
2. downstream search tab of the detail page -- when you open a node's
detail page, there's a tab that let's you see the upstream/downstream
dependencies; this page is also opened if you click the search icon that
pops out to the left of the expand button. the idea is that you can find
you node and click `Show Path on Graph`
3. right now, i opted to display _just_ the graph, sans filters and
other nodes, when you click the show path on graph in order to simply
this first iteration

https://github.com/user-attachments/assets/a701d670-fa51-4224-904b-c2d056d9fa5e

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

part of https://github.com/Recidiviz/recidiviz-data/issues/46966

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1a277a258a2f9c100e636acec4fe915c9b5d8585